refactor: InputView 분리
사용자의 입력을 받는 App.js 속 로직을 InputView로 분리
